Create root instance of Vue.js
When you create a vue instance, you will usually want to pass in a
configuration object as an argument. we've just added the el property
that tells Vue where to mount itself in the page.

You can assing to it a string (a CSS selector) or an HTML node object.
In our case, we've used the #app string, which is a CSS selector
referring to the element with the app ID.

Vue has dominion over the element it mounts on and any child node.