Create a data property in the vue instance
Any property of this data object will be available within our template.
To tell Vue where to bind this data, we can use mustache syntax, that
is, double curly brackets, for example, {{ myproperty }}. When Vue
instantiates, it compiles the template, replaces the mustache syntax
with the appropiate text, and updates the DOM to reflect this. This
process is called text interpolation