Work with some mock data
While we're developing, it'd be nice to work with some mock data so
that we can see how our completed page will look.

We've included sample/data.js in the project for this very reason.

Using global variables split over different script files in this way is
not an ideal practice. We'll only be doing this in the prototype,
though, and later we'll get this mock-listing sample from the server.