Add a header image
We've got a header image in our mock listing.

We're using a div rather than an img tag to help with positioning, we're
going to set our image as the background of the div with the header-img
class.

We use v-bind directive with the style argument to bind a value to the
style attribute.

headerImageStyle is an expression that evaluates to a CSS rule that sets
the background image to the correct URL.

We create headerImageStyle as data property. When binding to a style
attribute, you can use an object where the properties and values are
equivalent to the CSS properties and values.