Work with amenities
We add the v-for directive into the template and the expression we
assign to it will be amenity in amenities. The alias part of the
expression, that is amenity, will refer, throughout the loop sequence,
to each object in the array, starting with the first.

The second property of our amenity objects is icon. This is actually a
class relating to an icon in the Font Awesome icon font. We've installed
Font Awesome as an NPM module already, so add this to the head of the
page to now use it.