Allow the user to view the full text if they choose
We add a contracted class to the p tag that contains the about text
interpolation. The CSS rule for this class will restrict its height to
250 pixels and hide any text overflowing the element.

We also put a button after the p tag that the user can click to expand
the section to full height.

We need a way to remove the contracted class when the user clicks the
More button, using directives.

We create a contracted property and set its initial value to true.

Bind this class to an object. In the expression, the contracted property
is the name of the class to bound, the contracted value is a reference
to the data property evaluates to true, that class will be bound to the
element, and if it evaluates to false.

We want to remove the contracted class automatically when the user
clicks the More button. To do this job, we'll use the v-on directive,
which listens to DOM events with a click argument.

The value of the v-on directive can be an expression that assigns
contracted to false.

When the contracted value is changed by the click event, its set
method is triggered. The set method will set the new value, but will
also carry out a secondary task of informing Vue that a value has
changed and any part of the page relying on it may need to be
re-rendered.

We want to hide the More button as it's no longer needed. We can use the
v-if directive to achieve this in conjunction with the contracted
property.