Implement modal window
We'll use a modal window to allow the user to focus on a single item of
content.

The modal will give a properly scaled view of the header image so the
user can focus on the appearance of the lodgings without the distraction
of the rest of the page.

We add a Boolean data property that will represent the opened or closed
state of our modal.

We overlay a button labelled View Photos in the botton-left corner of
the header image.

Put the click listener on the wrapping div, the click event will we
captured regardless of whether the user clicks the button or the div
due to DOM event propagation.

The main modal div act as a container for the rest of the modal content
and as a background panel that will cover up the main window content.

we will dim the main window under a semi-transparent panel.