Load the mock data
We use the File facade that allows us to open files from disk. Then
convert the JSON file to a PHP array and once the data is an array, It
can be directly inserted into the database.

In order to insert the data, we'll use the DB facade and we'll call the
table method that allow us to query the database by chaining
constraints. We use the insert method, which accepts an array of column
names and values.

To execute the seeder we must call it from the DatabaseSeeder.php file.

We use the Artisan CLI to execute the seeder:
php artisan db:seed

Testint the mock data getting the first record:

DB::table('listings')->get()->first();