Add a $cast property to the Listing model
An Eloquent model can be given a $cast property to declare the data type
of any specific attribute. $cast is an array of key/values where the key
is the name of the attribute being cast, and the values is the data type
we want to cast to.

We'll cast the amenities attributes as Boolean.