Mock data dependency
$ cp ../vuebnb/sample/data.js ./resources/assets/js/

Currently, this file declares a global variable sample that is then
picked in the entry file. Let's make this file a module by replacing the
variable declaration with a ES2015 export default.

We can now import this module at the top of our entry file. Note that
Webpack can guess the file extension in an import statement so you can
omit the .js from data.js

While Laravel has opted to use CommonJS syntax for including modules,
that is require, we will use native ES module syntax, that is import.
This is because ES modules are making their way into the JavaScript
standard, and it's more consistent with the syntax used by Vue.