Get data from the backend into the frontend app
This commit include the following topics and changes:

We wrote the Vuebnb frontend app prototype using only ES5 syntax, as we
ran it directly in the browser without any build step.

Polyfills

The webpack babel plugin can transpile ES2015 syntax, but new API
methods require polyfilling. A polyfill is a script that is run it the
browser to cover an API or API method that may be missing.

Object.assign is a handy way of merging objects and would be useful in
our frontend app. We use Object.assign in our code, then add a polyfill
to ensure the code will run in older browsers.

We use Object.assign and simply merge the two objects. In practice, this
doesn't do anything different, it's just more succinct code.

To polyfill Object.assign we must install a new core-js dependency,
which is a library of polyfills for most new JavaScript APIs.

$ npm i --save-dev core-js

We add at the top of entry file app.js, this line to include the
Object.assign polyfill:

import "core/features/object/assign"

Let's modify on the web route file and change it to:

"/listing/{listing}"

Just like in our api/listing/{listing} route, the dynamic segment is
meant to match the ID for one of our mock data listings. If the dynamic
segment does not match a model, for example /listing/50 or
/listing/somestring, Laravel will abot the route and return a 404 HTTP
status.

Architecture

we can inject the model into the head of the page. That way, when the
Vue app loads, it will have immediate access to the model.

JSON

We'll now encode the entire model as JSON within the view. The JSON
format is good because it can be stored as a string and can be parsed by
both PHP and JavaScript.

We also had to wrap json_encode in another global function, addslashes.
This function will add backslashes before any character that needs to be
scaped. It's necessary to do this because the JavaScript JSON parser
doesn't know which quotes in the string are part of the JavaScript
syntax, and which are part of the JSON object.

We also had to use a different kind of Blade syntax for interpolation. A
feature of Blade is that statements withing double curly brackets {{}}
are automatically sent through PHP's htmlspecialchars function to
prevent XSS attacks. This will, unfortunately, invalidate our JSON
object. The solution is to use the alternative {!! !!} syntax, which
does not validate the contents. This is safe to do in this scenario
because we're sure we're not using any user-suplied content.

The safest way to share an object or function between two scripts is to
make it a property of the global window object.

window.vuebnb_listing_model

We remove the import sample from './data'; statement from the top of
app.js. We can also delete the sample data files as they won't be used
any further in the project:

$ rm resources/assets/js/data.js resources/assets/images/header.jpg

We put the transformation function into a file,
resources/assets/js/helpers.js.

The last thing to do is replace the hard-coded images' URLs in the
frontend app. We create a new add_image_urls method, which abstracts the
logic that was used in get_listing_api. Now both of the route-handling
methods will call this new method.

The final step is to use these image URLs from the backend instead of
the hard-coded URL. images is an array of URLs, we'll use the first
image as a default, that is, images[0].