Create our component image-carousel
This commit include the following topics until now:

We will use components to create an image carousel, which allows users
to peruse the different photos of a room listing.

Components are a tool for creating custom elements in Vue.js. When we
register a component, we define a template which renders as one or more
standard HTML elements.

The easiest way to register a component is to use the component API
method. The first argument is the name you want to give the component,
the second is the configuration object. The configuration will often
include a template property to declare the component's markup using a
string.

Once we've registered a component like this, we can use it withing our
project.

Component allow us to reuse JavaScript functionality. The configuration
object can not only include a template but can also include its own
state, just like the Vue instance. In fact, each component can be
thought of as a mini-instance of Vue with its own data, methods,
lifecycle hooks, and so on.

Tip:

A component can be referred to in your code by a kebab-case name such as
my-component, a PascalCase name such as MyComponent, or a camelCase
name such as myComponent. Vue sees these all as the same component.
However, in a DOM or string template, the component should always be
kebab-cased. Vue doesn't enforce this, but markup in the page gets
parsed by the browser before Vue gets to work with it, so it should
conform to W3C naming conventions ot parser may strip it out.

We register the component in our entry file. The template of this new
component will simply be the image tag we removed from the view, wrapped
in a div. We add this wrapping element, as a component templates must
have a single root element, and we'll soon be adding more elements
inside it.

Let's add a new data property, index, to our component, which will
dictate the current image being displayed. It will be initialized at 0
and the UI controls will later be able to increment or decrement the
value.

We will bind the image source to the array item at position index.

We use computed properties. These are properties we add to our Vue
configuration that can be thought of as reactive methods which are rerun
whenever a dependent value is changed.