Create a new component CarouselControl
This component will be presented as an arrowhead that floats over the
carousel and will respond to a user's click.

We register CarouselControl locally for the ImageCarousel component. The
CarouselControl template will render as an i tag, which is often used
for displaying icons. A good icon for carousel is the Font Awesome
chevron icon, which is an elegantly shaped arrowhead. Currently, we
don't have a way to distinguish between the left and right, so for now,
both instances will have a left-facing icon.