Bind root instance data property images to a prop
We can send data to a component through a custom HTML property know as
prop. We must also register this custom property in an array, props, in
the component's configuration.

A prop can be used just like any data property of the component: you can
interpolate it in the template, use it in methods and computed
properties, and so on. Think of prop data as being borrowed from another
component or instance - only the owner should change it.

TIP:
Props are proxied to the instance just like data properties, meaning you
can refer to a prop as this.myprop within that component's code.
Be sure to name your props uniquely to your data properties to avoid a
clash!

Since props must be declared in the template where the component is
used, prop data can only pass from a parent to a child. This is why you
shouldn't mutate a prop - since data flows down, the change will not be
reflected in the parent, and therefore you will have different versions
of what is meant to be the same bit of state.

We can reactively bind data to a component using the v-bind directive.
When the data changes in the parent, it will automatically flow down to
the child.

TIP:
Since the v-bind directive is used so commonly in templates, you can
omit the directive name as a shorthand: <div v-bind:title="title"> can
be shorthand to <div :title="title">

With props, we now have a mechanism for sending dynamic data from the
root instance down to a component. Let's bind the root instance data
property images to a prop, also called images, in our ImageCarousel
declaration.

The root instance will now be responsible for the state of the image
URLs, and the image carousel component will just be responsible for
displaying them.