Distinguish carousel controls
The CarouselControl component should have two possible states:
either left-pointing or right-pointing.

This state should not be internally determined, but instead
passed down from ImageCarousel. To do so, let's add a prop dir
to CarouselControl that will take a string value, and should be
either left or right.

With the dir prop, we can now bind the correct icon to the i
element. This is done with a computed property which appends
the prop's value to the string fa-chevron-, resulting in either
fa-chevron-left or fa-chevron-right.