Change carousel images
Returning to CarouselControl, let's respond to user's click by using the
v-on directive and triggering a method, clicked. This method will, in
turn emit a custom event, change-image, which will include a payload of
either -1 or 1, depending on whether the state of the component is left
or right.

TIP:
Just like with v-bind, there is a shorthand for v-on as well. Symply
replace v-on: with @; for instance, <div @click="handler"></div> is the
equivalent of <div v-on:click="handler"></div>

ImageCarousel will now need to listen for the change-image event via the
v-on directive. The event will be handled by a method changeImage which
will have a single parameter, val, reflecting the payload being sent in
the event. The method will then use val to step the value of index,
ensuring it loops to the start or end if it exceeds the bounds of the
array it indexes.