Refactoring ImageCarousel vue
The following topics is covered:

Single-file components (SFCs) are files with a .vue extension that
contain the complete definition of a single component and can be
imported into your Vue.js app. SFCs make it simple to create and use
components.

SFCs are similar to HTML files but have three root elements:

template
script
style

The component definition goes inside the script tag and will be exactly
like any other component definition except:

It will export an ES module
It will not need a template property (or a render function; more on that
later)

The component's template will be declared as HTML markup inside the
template tag.

Transformation

To use a single-file component in your app, you simply import it like it
were and ES module. The .vue file is not a valid JavaScript module file,
however. Just like we use the Webpack Babel plugin to transpile our
ES2015 code into ES5 code, we must use Vue Loader to transform .vue
files into JavaScript modules.

Vue Loader is already configured by default with Laravel Mix, so there's
nothing further we need to do in this project; any SFC's we import will
just work!

We create a new directory, then we will create the .vue files:

$ mkdir resources/assets/components
$ touch resources/assets/components/ImageCarousel.vue
$ touch resources/assets/components/CarouselControl.vue

we move the template string into the template tag, and the component
definition into the script tag. The component definition must be
exported as a module.

Now we can import this file into our app and register it locally in the
root instance. As mentioned, Vue is able to automatically switch between
kebab-case component names and Pascal-case ones. This means we can use
the object shorthand syntax inside the component configuration and Vue
will correctly resolve it.

SFCs allow us to add style to a component, helping to better organize
our CSS code. Let's move the CSS rules we created for the image carousel
into this new SFC's style tag.

When the project builds, you should find it still appears the same. The
interesting thing, though, is where the CSS has ended up in the build.
If you check public/css/style.css, you'll find it's not there.

To use it, Webpack's bootstrapping code will inline this CSS string into
the head of the document when the app runs.

Inlining CSS is actually the default behavior of Vue Loader. However, we
can override this and get Webpack to write SFC styles to their own file.
Add the following to the bottom of the Mix configuration.

Now an additional file, public/css/vue-style.css, will be outputted in
the build.

We'll need to load this new file in our view, after the main style
sheet.