Use ref to call the modal
ref is a especial property that allows you to directly reference a child
component's data. To use it, declare the ref property and assign it a
unique value, such as imagemodal.

Now the root instance has access to this specific ModalWindow
component's data via the $ref object. This means we can change the value
of modalOpen inside a root instance method, just like we could from
within ModalWindow.

TIP
It is anti-pattern to use ref when the normal methods of interacting
with a component, props and events, are sufficient. ref is usually only
required for communicating with elements that fall outside of the normal
flow of a page, as a modal window does.