Create a new FeatureList component file
We refactoring the list of amenities and prices, and abstract on a
component. These lists have a similar purpose and structure, so it makes
sense that we create a single component for both.

We observe that the main difference between the two lists is inside the
<div class="content">...</div> section, as the data being displayed in
each list has a slightly different structure. The ameninities have a
icon and title whereas the prices have a title and value. We'll use a
slot in this section to allow the parent to customize the content for
each.

We'll move the markup for one of the lists in, using a slot to replace
the list content. We'll also add a prop for the title and move in any
list-related CSS.

We import the FeatureList component into app.js, and add it to the
locally registered components.