Use scoped slot for prices and amenities lists
Scoped slots allow you to pass a template to the slot instead of passing
a rendered element. Whe this template is declared in the parent, it will
have access to any props supplied in the child.

We want to be able to pass the list items array in as a prop and get the
FeatureList component to iterate them. That way, FeatureList is taking
ownership of any repeated functionality. The parent will then provide a
template to define how each list item should display.

- Add items to the props array in the configuration object
- items will be array which we iterate inside the <div class="content">
section
- In the loop, item is an alias to any particular list item. We can
create a slot and bind that list item to the slot using v-bind="item".
This will bind the properties of an entire object to the element. This
useful as the amenities and prices objects have a different
properties and we don't have to specify them.

Return on the view and we make the following:

- Declare a template element inside the FeatureList declaration.
- The template must include the slot-scope property to which we assign
and alias amenity. This alias allow us to access the scoped props.
- Inside the template, we can use exactly the same markup we had before
to display our amenity list items.