Refactor the main template as single-file component
Firstly, we're going to talk about to Virutal DOM, in order for Vue to
be able to render a component to the page, it will first transform the
template string into a JavaScript object using an internal template
compiler library.

a-component -> template-compiler -> JavaScript-object

Once the template has been compiled, any state or directives can easily
be applied.

After that, Vue can interface with the DOM API to synchronize the page
with the state of the component.

Render functions

Render functions are more efficient because they don't require Vue to
first compile the template string. The downside, though, is that writing
a render function is not as easy or expressive as markup syntax, and,
once you get a large template, will be difficult to work with.

Vue Loader

It would be great if we could create HTML markup template in
development, then get Vue's template compiler to turn them into render
functions as part of the build step. This is exactly what happens to
single-file components when webpack transform them via Vue Loader.

Refacotr the main template

The template for the app's root intance is the content within the #app
element in the app view. A DOM template like this requires the Vue
template compiler, just like any string template does.

If we were able to abstract this DOM template into an SFC as well, it
would mean all our frontend app templates would be built as render
functions and would not need to invoke the template compiler at runtime.

Let's create a new SFC for the main template and call it ListingPage, as
this part of the app is our listing page.

We'll move the main template, root configuration and any relevant CSS
into this component.

- we need to put the template inside the wrapping div as component must
have a single root element.
- we can now remove the @ scapes as this file won't be processed by
Blade.
- The component is now adjacent to the other components we created, so
be sure to change the relative paths of the imports.

Mounting the root-level component with a render function

We'll keep the mount element empty and declare Listing with a render
function inside our root instance.

Now that Vuebnb is no longer using string or DOM templates, we don't
need the template compiler functionality anymore. There's a special
build of Vue we can use which doesn't include it!.

Vue.js builds

Theere are a number of different environments and use cases for running
Vue.js. In one project, you might load Vue directly in the browser, in
another you may load it on a Node.js server for the purpose of server
rendering. As such, there are different builds of Vue provided so you
can choose the most suitable one.

Looking in the dist folder of the Vue NPM package,we can seee eight
different Vue.js builds.

                 UMD               CommonJS               ES Module
Full          - vue.js         - vue.common.js         -  vue.esm.js
Runtime-only  - vue.runtime.js - vue.runtime.common.js - vue.runtime.esn
Full          - vue.min.js     -                       -
(production)
Runtime-only
(production)  - vue.runtime    -                       -
                .min.js

Module System

The main things you need to know about UMD is that it's yet another
module pattern, and that it works well in a browser. UMD is the choice
if you are directly linking to Vue in a script tag.

Production builds

A production build is used in a deployed app. It has been minified, and
way and any warnings, comments or other development options are turned
off or stripped out. The point is to make the build as small and secure
as possible, which is what you'd want in production.

Note that there is only a UMD version of the production build as only
UMD runs directly in a browser. CommonJs and ES Module are to be used in
conjunction with a build tool, like Webpack, which provides its own
production processing.

Full build vs runtime-only

Vue indcludes a template compiler for converting any string or DOM
templates to render functions at runtime. The full build includes the
template compiler and is what you would normally use. However, if you've
already transformed your templates into render functions in development,
you can use the runtime-only build, which drops the compiler and is
about 30% smaller!.

Selecting a build

A good build for Vuebnb is vue.runtime.esm.js since we're using Webpack
and we don't need the template compiler. We could also use
vue.runtime.common.js, but that wouldn't be consistent with our use of
ES modules elsewhere in the project. In practice, though, there is no
difference as Webpack will process them in the same way.

We inlucde Vue at the top of our entry file with the statement import
Vue from 'vue'. The last 'vue' is an alias to the Vue build that Webpack
resolves when it runs. Currently, this alias is defined withing the
default Mix configuration and is set to the build vue.common.js. We can
override that configuration by adding the following to the bottom of our
webpack.mix.js file.

After a new build, we should expect to see a smaller bundle size due to
the template compiler being removed.

Keep in mind that without the template compiler we can no longer provide
string template for our components. Doing so will cause an error at
runtime.