Create routes
The most basic configuration for Vue Router is to provide a routes
array, which maps URLs to the corresponding page components. This array
will contain objects with at least two properties: path and component.

We're only going to have two routes in our app, one for our home page
and one for our listing page.

The path for our ListingPage component contains a dynamic segment
:listing so that this route will match paths including /listing/1,
/listing/2 ... listing/whatever

NOTE

There are two modes for Vue Router: hash mode and history mode.
Hash mode uses the URL hash to simulate a full URL so that the page
won't be reloaded when the hash changes. History mode has real URL and
leverages the history.pushState API to change the URL causing a page
reload. The only downside to history mode is that URLs outside of the
app, such as /some/weird/path, can't be handled by Vue and must handled
by the server. So we'll use history mode for Vuebnb