Create a App.vue component to be a root component
We need to declare a RouterView component somewhere in our page
template.

We'll slightly restructure our app to do this. As it is, the ListingPage
component is the root component of the app, as it is at the top of the
component hierarchy and loads all other components that we use.

Since we want the router to switch between ListingPage and HomePage
based on the URL, we need another component to be above ListingPage in
the hierarchy and handle this work. This component it'll call App.

Let's create the App component file

$ touch resources/assets/component/App.vue

The root instance of Vue should render this to the page when it loads,
instead of ListingPage.

The following is the content of the App component. I've added the
special RouterView component into the template, which is the outlet
where either the HomePage or ListingPage component will render.

You'll also notice I've moved the toolbar from app.blade.php into the
template of App. This is so the toolbar is in the domain of Vue; before
it was outside of the mount point and therefore untouchable by Vue. I've
donde this so that later we can make the main logo link to the home page
using RouterLink, as this is a convention for most websites. I've moved
any toolbar related CSS into the style element as well.

If you now navigate the browser to a URL like /listing/1, you'll see
everything looks the same as it did before. However, if you look at Vue
Devtools, you'll see the component hierarchy has changed, reflecting the
addition of the App component.

There's also an indicator, which tells us that the ListingPage component
is the active page component for Vue Router