Create a HomePage component
Let's create a new component, HomePage

$ touch resources/assets/components/HomePage.vue

Let's add placeholder markup to the component before we set it up
properly.

We import the component in the router file, and uncomment the route
where it's used.

Let's now create a link to our home page in the toolbar by using the
RouterLink component. This component is like an enhanced a tag.

We give routes a name property, we can simply use the to prop rather
than having to supply an href. Vue resolve this to the correct URL on
render.

Let's add name properties to our routes.

We have to modify our CSS now since we now have another tag wrapped
around our logo.

Let's now open listing page, such as /listing/1. If you inspect the DOM,
we'll see that our toolbar now has a new a tag indide it with a
correctly resolved link back to the home page.

If we click that link, you'll be taken to the home page! Remember, the
page hasn't actually changed; Vue router simply swapped ListingPage for
HomePage within RouterView, and also updated the browser URL via the
history.pushState API.