Add a server-side route
Add a server-side route for the home page so that we can load our app
from the root path. This new route will point to a get_home_web method
in our ListingController class.

The get_home_web method returns the app view. The app view includes a
template variable model which we use to pass through the initial
application state. For now, just assign an empty array as a placeholder.

When the Vue app is bootstrapped, Vue Router will check the URL value
and, seeing that the path is /, will load the HomePage component inside
the RouterView outlet for the first rendering of the app.