Refactor some variables and restructuring some methods
This change will ensure that the code semantics reflect the changing
requirements and keep our code readable and easy to understand.

We rename our template variable from $model to the more general $data.

We're now going to abstract any common logic from our listing route
methods into a new helper method called get_listing. In this helper
method, we will nest the Listing model inside a Laravel Collection under
the listing key. Collection is an array-like wrapper for Eloquent models
that offers a bunch of handy methods that we'll be putting to use
shortly. get_listing will include logic from the add_image_urls helper
method, which can now safely be deleted.

We'll also need to reflect the change to our template variable when we
call the view method.

Finally, we'll need to update our ListingPage component to reflect the
new name and structure of the server data we're injecting.