Add a thumbnail to the listing
Each mock listing has a thumbnail version of the first image, which can
be used for the listing summary. The thumbnail is much smaller than the
image we use for the header of the listing page and is ideal for the
listing summaries on the home page. The URL for the thumbnail is
public/images/{x}/Image_1_thumb.jpg where {x} is the ID of the listing.