Group the listing for country
We've parsed our injected data, let's modify the object so the listing
are grouped by country. We can easily create a function is always
explicitly named, for example, No. 51, Hanzhong Street, Wanhua District,
Taipei City, Taiwan 108.

We've supplied one in the helpers module called groupByCountry which can
be imported at the top of the component configuration.