Create ListingSummary component vue
To begin with, clear out the existing content of the component and
replace it with a div. This div will feature a v-for directive to
iterate through each of our listing groups. Since listing_groups is and
object with key/value pairs, we'll give our v-for two aliases: group and
country, which are the value and key of each object item respectively.

We'll interpolate country inside a heading group.

We'll create a separate component, ListingSummary, for displaying them.

$ touch resources/assets/components/ListingSummary.vue

Let's declare ListingSummary withing our HomePage template. We'll again
a v-for directive to iterate group, an array, creating a new instance of
ListingSummary for each number. The data for each member will be bound
to a single prop, listing.

We gave each listing summary a fixed width/height so that we could
display them in a neat grid. Currently, they're displaying in one tall
column, so let's add some CSS flex rules to the HomePage component to
get the summaries into rows.

We'll add a class listing-summary-group to the element that wraps the
summaries. We'll also add a class home-container to the root div to
constrain the width of the page and center the content.

Finally, we'll need to add a rule to prevent the listing from forcing
the edge of the document to exceed the viewport. We add this to the main
CSS file.