Add path to the model
We add a property to the data injected into the head of our view. To do
this, we'll add a helper function called add_meta_data which will add
the path.

Note that the path of the current route can be determined by the Request
object. This object can be declared as the last argument of any
route-handling functions and is provided in each request by the service
container.