Let's add beforeRouteEnter to the HomePage component
Firstly, move any logic for retrieven data from the document head into
the hook. We then check the path property of the data to see if it
matches the current route. If so, we call next and pass a callback
function that applies the data to the component's instance. If not,
we'll need to use AJAX to get the right data.

We've added listing_groups as a data property. Before, we were applying
our data to the component instance as it was created. Now, we're
applying the data after the component is created. To set up reactivity
data, Vue must know the names of the data properties, so we initialize
with an empty value and update it when the data needed is available.