Add home API endpoint
We'll now implement the AJAX functionality. Before we do, though, we
need to add a home page endpoint to our web service.

- Let's firt add the home API route.

Looking now at the ListingController class, we'll abstract the bulk of
the logic from get_home_web into a new function, get_listing_summaries.
We'll then use this function in the get_home_api method and return a
JSON responde.