Use axios
To perform AJAX request to the web service, we'll use the Axios HTTP
Client, which is included with Laravel's default frontend code. Axios
has a very simple API allowing us to make request to a GET URL.

Axios is a promise-based library, so in order to retrive the response,
you can simply chain a then callback.

As the Axios NPM package is already installed, we can go ahead and
import the HomePage component. We can then use it to perform the request
to the home API endpoint, /api/. In the then callback, we apply the
returned data to the component instance exactly as we did with the
inlined model.

And with that, we can now navigate to the home page in two ways, either
via the address bar, or by going from a link from the listing page.
Either way, we get the right data!.