Move the solution to a mixin
Let's generalize the solution for getting the right data to the home
page so that we can use it on the listing page as well. To do this,
we'll move Axios and the beforeRouteEnter hook from the HomePage
component into a mixin that can then be added to both page components.

$ touch resources/assets/js/route-mixin.js

At the same time, let's improve the code by removing the repetition of
the next function call. To do this, we'll create a new method, getData,
which will be responsible for figuring out where to get the right data
for the page and also for getting it. Note that this method will be
asynchronous since it may need to wait for AJAX to resolve, so it will
return a Promise rather than an actual value.

This promise is then resolved within the navitgation guard.

NOTE

We don't need a polyfill for promise as that is already supplied in the
Axios library.