Call to the assignData method
You'll notice that within the next callback we call a method on the
subject component called assignData, passing the data object as an
argument. We'll need to implement the assignData method in any component
that uses this mixin. We do it this way so that the component can
process the data, if necessary, before it is applied to the component
instance. For example, the ListingPage component must process the data
via the populateAmenitiesAndPrices helper function.

We'll also need to add assignData to the HomePage component.