Link to the listing page
Each of our ListingSummary instaces represents a single listing, and
should therefore be clickable link to the page for that listing. Let's
use the RouteLink component to achieve this. Note that the object we
bind to the to prop includes the name of the routes as well as a params
object which includes a value for the dynamic segment of the route, the
listing ID.

The listing summaries will now be links. Clicking from one to the
listing page.

We can see that the AJAX call to the listing API was successful and
returned the data we wanted. If we also look at the Vue Devtools tab, as
well as the Dev Tools console, we can see the correct data in our
component instance. The problem is that we now have an unhandled 404
error for the header image.

The reason for this is that the component's first render ocurs before
the callback in the next hooks is called. This means that the
initialization values for the component data are used in the first
render.

In the HeaderImage declaration, we bind the first image like this:
:image-url="images[0]". Since the array is initially empty, this will be
an undefined value and results in the unhandled error.

The explanation is complex, but the fix is easy: just add a v-for to
header-image, ensuring it won't render until valid data is available.